PCB Renaming
Since the name of the projects included "PCB, [NAME PANEL]",  anytime manufacturing files or anything were created, it named all of the file within the project the same.  This includes the BOM and CPL files which couldn't be uploaded to JLC due to the fact that they had a comma in the name.

Also, having "PCB" included in the name of the PCB, under a folder called "PCBs" with in the Electrical CAD folder, seemed kind of ..... redundant.

Looks cleaner and easier to navigate by removing PCB from the front.  Each project is being saved as a new project with a new name, which automatically changes the name of each file within its subdirectory (with a few minor exceptions).  The files that cannot be changed automatically will be changed manually to follow suit.